deep_research_agent/requirements.txt
streamlit
duckduckgo-search
beautifulsoup4
requests
sentence-transformers
faiss-cpu
transformers
torch
fpdf
duckduckgo-search